refactor: remove alipay related strings
