+++ server/home/home-service/src/main/java/io/holoinsight/server/home/biz/plugin/core/MysqlPlugin.java
@@ -6,7 +6,7 @@ package io.holoinsight.server.home.biz.plugin.core;
 
 import io.holoinsight.server.home.biz.plugin.model.PluginModel;
 import io.holoinsight.server.home.dal.model.dto.IntegrationPluginDTO;
-import com.alipay.cloudmonitor.registry.integration.model.mysql.MysqlTask;
+import io.holoinsight.server.registry.model.integration.mysql.MysqlTask;
 import io.holoinsight.server.common.J;
 import com.google.gson.reflect.TypeToken;
 import org.springframework.stereotype.Component;
@@ -20,7 +20,7 @@ import java.util.List;
  * @version 1.0: MysqlPlugin.java, v 0.1 2022年11月22日 上午11:14 jinsong.yjs Exp $
  */
 @Component
-@PluginModel(name = "com.alipay.holoinsight.plugin.MysqlPlugin", version = "1")
+@PluginModel(name = "io.holoinsight.plugin.MysqlPlugin", version = "1")
 public class MysqlPlugin extends AbstractCentralIntegrationPlugin<MysqlPlugin> {
 
   public MysqlTask mysqlTask;
