+++ server/registry/registry-model/src/main/java/io/holoinsight/server/registry/model/integration/CentralIntegrationTask.java
@@ -1,7 +1,7 @@
 /*
  * Copyright 2022 Holoinsight Project Authors. Licensed under Apache-2.0.
  */
-package com.alipay.cloudmonitor.registry.integration.model;
+package io.holoinsight.server.registry.model.integration;
 
 import lombok.Data;
