+++ server/registry/registry-model/src/main/java/io/holoinsight/server/registry/model/integration/alicloud/AlicloudConf.java
@@ -1,7 +1,7 @@
 /*
  * Copyright 2022 Holoinsight Project Authors. Licensed under Apache-2.0.
  */
-package com.alipay.cloudmonitor.registry.integration.model.alicloud;
+package io.holoinsight.server.registry.model.integration.alicloud;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
