+++ server/registry/registry-model/src/main/java/io/holoinsight/server/registry/model/integration/springboot/SpringBootTask.java
@@ -1,9 +1,9 @@
 /*
  * Copyright 2022 Holoinsight Project Authors. Licensed under Apache-2.0.
  */
-package com.alipay.cloudmonitor.registry.integration.model.springboot;
+package io.holoinsight.server.registry.model.integration.springboot;
 
-import com.alipay.cloudmonitor.registry.integration.model.LocalIntegrationTask;
+import io.holoinsight.server.registry.model.integration.LocalIntegrationTask;
 import lombok.Data;
 import lombok.NoArgsConstructor;
